refactor(assay): extract derive.rs module
